requirements.txt
# For generic git repo access
gitpython
PyGithub
streamlit
streamlit-keyup
fastapi
uvicorn[standard]
python-dotenv
anthropic
pydantic
